pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "twinbox-core"
version = "0.1.0"
description = "Lightweight email intelligence for OpenClaw"
requires-python = ">=3.11"
dependencies = ["PyYAML>=6.0"]

[project.scripts]
twinbox = "twinbox_core.cli:main"

[tool.setuptools.packages.find]
include = ["twinbox_core*"]
